Refactor extensions code
